requirements.txt
art
kurigram
aiofiles
#gcrypto
requests
py-tgcalls
pytz
google
numpy
python-dotenv
plyer
SQLAlchemy[asyncio]
asyncpg
aiosqlite
